Added linter, prettier and fixed base project configurations
