+++ wbooks/__tests__/App-test.tsx
@@ -4,10 +4,11 @@
 
 import 'react-native';
 import React from 'react';
+import renderer from 'react-test-renderer';
+
 import App from '../App';
 
 // Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
   renderer.create(<App />);
