+++ wbooks/index.d.ts
@@ -1,0 +1,2 @@
+declare module '*.png';
+declare module '*.json';
