+++ wbooks/metro.config.js
@@ -7,11 +7,12 @@
 
 module.exports = {
   transformer: {
+    // eslint-disable-next-line require-await
     getTransformOptions: async () => ({
       transform: {
         experimentalImportSupport: false,
-        inlineRequires: false,
-      },
-    }),
-  },
+        inlineRequires: false
+      }
+    })
+  }
 };
