+++ wbooks/package.json
@@ -1,13 +1,17 @@
 {
   "name": "wbooks",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
-    "start": "react-native start",
+    "android": "npx react-native run-android",
+    "android:build": "cd android/ && ./gradlew clean assembleRelease",
+    "android:clean": "cd android/ && ./gradlew clean",
+    "ios": "npx react-native run-ios",
+    "start": "npx react-native start",
     "test": "jest",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
+    "lint-fix": "eslint . --fix --ext .js,.jsx,.ts,.tsx",
+    "check-types": "tsc"
   },
   "dependencies": {
     "react": "16.13.1",
@@ -22,10 +26,28 @@
     "@types/react-test-renderer": "^16.9.2",
     "@typescript-eslint/eslint-plugin": "^2.27.0",
     "@typescript-eslint/parser": "^2.27.0",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
+    "babel-plugin-import-glob": "^2.0.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "eslint": "^6.5.1",
+    "eslint-config-airbnb": "^18.1.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-config-wolox-react-native": "^1.0.1",
+    "eslint-import-resolver-babel-module": "^5.1.2",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-flowtype": "^4.7.0",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-react": "^7.19.0",
+    "eslint-plugin-react-hooks": "^4.0.0",
+    "eslint-plugin-react-native": "^3.8.1",
+    "husky": "^4.2.5",
     "jest": "^25.1.0",
     "metro-react-native-babel-preset": "^0.59.0",
+    "prettier": "^2.0.5",
+    "prettier-eslint": "^9.0.1",
     "react-test-renderer": "16.13.1",
     "typescript": "^3.8.3"
   },
@@ -39,5 +61,10 @@
       "json",
       "node"
     ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "yarn run lint && yarn run check-types"
+    }
   }
 }
